parser/requirements.txt
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
fuzzywuzzy==0.18.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.1
idna==3.6
Levenshtein==0.24.0
loguru==0.7.2
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PySocks==1.7.1
python-Levenshtein==0.24.0
rapidfuzz==3.6.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
soupsieve==2.5
StrEnum==0.4.15
urllib3==2.2.0
